refator: moving bitrate logic and dvr to single avfountation extension files
