+++ Sources/Clappr/Classes/Plugin/Playback/AVFoundationPlayback.swift
@@ -611,73 +611,4 @@ open class AVFoundationPlayback: Playback {
             view.accessibilityIdentifier = "AVFoundationPlaybackError"
         }
     }
-    
-    open func getBitrate() -> Double? {
-        guard let logEvent = lastLogEvent() else { return nil }
-        if (logEvent.segmentsDownloadedDuration ) > 0 {
-            return logEvent.indicatedBitrate
-        }
-        return logEvent.observedBitrate
-    }
-    
-    open func getAvgBitrate() -> Double? {
-        guard let logEvent = lastLogEvent() else { return nil }
-        return logEvent.averageVideoBitrate
-    }
-    
-    private func lastLogEvent() -> AVPlayerItemAccessLogEvent? {
-        return player?.currentItem?.accessLog()?.events.last
-    }
-}
-
-// MARK: - DVR
-extension AVFoundationPlayback {
-    open override var minDvrSize: Double {
-        return options[kMinDvrSize] as? Double ?? 60.0
-    }
-
-    open override var isDvrInUse: Bool {
-        if state == .paused && isDvrAvailable { return true }
-        guard let end = dvrWindowEnd, playbackType == .live else { return false }
-        guard let currentTime = player?.currentTime().seconds else { return false }
-        return end - liveHeadTolerance > currentTime
-    }
-
-    open override var isDvrAvailable: Bool {
-        guard playbackType == .live else { return false }
-        return duration >= minDvrSize
-    }
-
-    open override var currentDate: Date? {
-        return player?.currentItem?.currentDate()
-    }
-
-    open override var seekableTimeRanges: [NSValue] {
-        guard let ranges = player?.currentItem?.seekableTimeRanges else { return [] }
-        return ranges
-    }
-
-    open override var loadedTimeRanges: [NSValue] {
-        guard let ranges = player?.currentItem?.loadedTimeRanges else { return [] }
-        return ranges
-    }
-
-    open override var epochDvrWindowStart: TimeInterval {
-        guard let currentDate = currentDate else { return 0 }
-        return currentDate.timeIntervalSince1970 - position
-    }
-
-    private var dvrWindowStart: Double? {
-        guard let end = dvrWindowEnd, isDvrAvailable, playbackType == .live else { return nil }
-        return end - duration
-    }
-
-    private var dvrWindowEnd: Double? {
-        guard isDvrAvailable, playbackType == .live else { return nil }
-        return seekableTimeRanges.max { rangeA, rangeB in rangeA.timeRangeValue.end.seconds < rangeB.timeRangeValue.end.seconds }?.timeRangeValue.end.seconds
-    }
-
-    fileprivate var liveHeadTolerance: Double {
-        return 5
-    }
 }
